[ALPS05377541] PMIC: Fix coverity scan
Avoid evaluates to an address that could be at negative offset of an array

MTK-Commit-Id: d5b1ba27f6630afc5b2381c4c0ab330849bf6a71

CR-Id: ALPS05377541
Feature: [Module]PMIC
(cherry picked from commit ffd5f532287baf22b3bf9f787a3db629caca3e02)
(cherry picked from commit 365652312f4f1e4edfca8a9c001f935a4217d1f7)

+++ drivers/misc/mediatek/pmic/mt6359p/v1/pmic_throttling_dlpt.c
@@ -630,7 +630,7 @@ void register_battery_percent_notify_ext(
 {
 	PMICLOG("[%s] start\n", __func__);
 
-	bpcb_tb_ext[prio_val].bpcb = battery_percent_callback;
+	bpcb_tb_ext[(unsigned int)prio_val].bpcb = battery_percent_callback;
 
 	pr_info("[%s] prio_val=%d\n", __func__, prio_val);
